# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Apexy VPN API",
    "version": "1",
    "description": "Generated from server/api route definitions. See docs/BACKEND_API.md."
  },
  "servers": [
    {
      "url": "https://api.apexyvpn.example"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT (EdDSA, 15 min)"
      },
      "nodeToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Per-node token (VPN nodes only)"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/v1/subscription": {
      "get": {
        "tags": [
          "subscription"
        ],
        "summary": "Current plan and billing state.",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "none",
                        "incomplete",
                        "trialing",
                        "active",
                        "past_due",
                        "canceled",
                        "expired"
                      ]
                    },
                    "plan": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            },
                            "period": {
                              "type": "string",
                              "enum": [
                                "trial",
                                "month",
                                "year"
                              ]
                            },
                            "priceCents": {
                              "type": "number"
                            },
                            "currency": {
                              "type": "string"
                            },
                            "deviceLimit": {
                              "type": "number"
                            }
                          },
                          "required": [
                            "id",
                            "name",
                            "period",
                            "priceCents",
                            "currency",
                            "deviceLimit"
                          ]
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "currentPeriodStart": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "currentPeriodEnd": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "cancelAtPeriodEnd": {
                      "type": "boolean"
                    },
                    "provider": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "paymentMethod": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "brand": {
                              "type": "string"
                            },
                            "last4": {
                              "type": "string"
                            },
                            "expMonth": {
                              "type": "number"
                            },
                            "expYear": {
                              "type": "number"
                            }
                          },
                          "required": [
                            "brand",
                            "last4",
                            "expMonth",
                            "expYear"
                          ]
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "devicesUsed": {
                      "type": "number"
                    },
                    "deviceLimit": {
                      "type": [
                        "number",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "plan",
                    "currentPeriodStart",
                    "currentPeriodEnd",
                    "cancelAtPeriodEnd",
                    "provider",
                    "paymentMethod",
                    "devicesUsed",
                    "deviceLimit"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/subscription/plans": {
      "get": {
        "tags": [
          "subscription"
        ],
        "summary": "Plans on offer.",
        "security": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "period": {
                        "type": "string",
                        "enum": [
                          "trial",
                          "month",
                          "year"
                        ]
                      },
                      "priceCents": {
                        "type": "number"
                      },
                      "currency": {
                        "type": "string"
                      },
                      "deviceLimit": {
                        "type": "number"
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "period",
                      "priceCents",
                      "currency",
                      "deviceLimit"
                    ]
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/subscription/checkout": {
      "post": {
        "tags": [
          "subscription"
        ],
        "summary": "Start (or change to) a plan. Returns a hosted checkout URL, or `activated` when the provider needs no payment step.",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "planId": {
                    "type": "string"
                  },
                  "returnTo": {
                    "default": "app",
                    "type": "string",
                    "enum": [
                      "app",
                      "web"
                    ]
                  }
                },
                "required": [
                  "planId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "kind": {
                      "type": "string",
                      "enum": [
                        "redirect",
                        "activated"
                      ]
                    },
                    "url": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "kind"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/subscription/cancel": {
      "post": {
        "tags": [
          "subscription"
        ],
        "summary": "Cancel at the end of the paid period. Access continues until then; nothing is refunded or cut short.",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "none",
                        "incomplete",
                        "trialing",
                        "active",
                        "past_due",
                        "canceled",
                        "expired"
                      ]
                    },
                    "plan": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            },
                            "period": {
                              "type": "string",
                              "enum": [
                                "trial",
                                "month",
                                "year"
                              ]
                            },
                            "priceCents": {
                              "type": "number"
                            },
                            "currency": {
                              "type": "string"
                            },
                            "deviceLimit": {
                              "type": "number"
                            }
                          },
                          "required": [
                            "id",
                            "name",
                            "period",
                            "priceCents",
                            "currency",
                            "deviceLimit"
                          ]
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "currentPeriodStart": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "currentPeriodEnd": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "cancelAtPeriodEnd": {
                      "type": "boolean"
                    },
                    "provider": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "paymentMethod": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "brand": {
                              "type": "string"
                            },
                            "last4": {
                              "type": "string"
                            },
                            "expMonth": {
                              "type": "number"
                            },
                            "expYear": {
                              "type": "number"
                            }
                          },
                          "required": [
                            "brand",
                            "last4",
                            "expMonth",
                            "expYear"
                          ]
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "devicesUsed": {
                      "type": "number"
                    },
                    "deviceLimit": {
                      "type": [
                        "number",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "plan",
                    "currentPeriodStart",
                    "currentPeriodEnd",
                    "cancelAtPeriodEnd",
                    "provider",
                    "paymentMethod",
                    "devicesUsed",
                    "deviceLimit"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/subscription/resume": {
      "post": {
        "tags": [
          "subscription"
        ],
        "summary": "Undo a pending cancellation.",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "none",
                        "incomplete",
                        "trialing",
                        "active",
                        "past_due",
                        "canceled",
                        "expired"
                      ]
                    },
                    "plan": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            },
                            "period": {
                              "type": "string",
                              "enum": [
                                "trial",
                                "month",
                                "year"
                              ]
                            },
                            "priceCents": {
                              "type": "number"
                            },
                            "currency": {
                              "type": "string"
                            },
                            "deviceLimit": {
                              "type": "number"
                            }
                          },
                          "required": [
                            "id",
                            "name",
                            "period",
                            "priceCents",
                            "currency",
                            "deviceLimit"
                          ]
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "currentPeriodStart": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "currentPeriodEnd": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "cancelAtPeriodEnd": {
                      "type": "boolean"
                    },
                    "provider": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "paymentMethod": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "brand": {
                              "type": "string"
                            },
                            "last4": {
                              "type": "string"
                            },
                            "expMonth": {
                              "type": "number"
                            },
                            "expYear": {
                              "type": "number"
                            }
                          },
                          "required": [
                            "brand",
                            "last4",
                            "expMonth",
                            "expYear"
                          ]
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "devicesUsed": {
                      "type": "number"
                    },
                    "deviceLimit": {
                      "type": [
                        "number",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "plan",
                    "currentPeriodStart",
                    "currentPeriodEnd",
                    "cancelAtPeriodEnd",
                    "provider",
                    "paymentMethod",
                    "devicesUsed",
                    "deviceLimit"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/subscription/invoices": {
      "get": {
        "tags": [
          "subscription"
        ],
        "summary": "Invoice history.",
        "security": [
          {
            "bearer": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "number": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      },
                      "amountCents": {
                        "type": "number"
                      },
                      "currency": {
                        "type": "string"
                      },
                      "status": {
                        "type": "string"
                      },
                      "issuedAt": {
                        "type": "string"
                      },
                      "paidAt": {
                        "type": [
                          "string",
                          "null"
                        ]
                      }
                    },
                    "required": [
                      "id",
                      "number",
                      "description",
                      "amountCents",
                      "currency",
                      "status",
                      "issuedAt",
                      "paidAt"
                    ]
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}